chore: remove unused libraries and files, and updated readme.md
