+++ frontend/README.md
@@ -2,8 +2,8 @@
 
 ## Setting up your local environment
 
-- Ensure that you have `node v18+`, and `npm`.
-- Clone the repository in your machine. You can view detailed instructions [here](https://docs.google.com/spreadsheets/d/1fs1PiRtqxIcetbEjcDo6Adia3vk7rxpiOeCc2Egtb3s/edit#gid=2016470823).
+- Ensure that you have `node 22.15+` installed.
+- Clone the repository in your machine.
 - Open the repository in your IDE
 - Open up your IDE's terminal. Make sure you're at the root folder of the project. Enter `cd frontend` in the terminal
   ![image](https://github.com/SPARCS-UP-Mindanao/SPARCS-Event-Platform/assets/85269524/adcdd520-d032-4380-8a85-81bbf39805c2)
@@ -35,22 +35,6 @@ With the help of `tsconfig.json`, we are able to use aliases for our directories
 We can simply use `@/{folderInSRC}` for accessing folders or files under the `frontend/src` directory.
 This makes it easy for us as we don't need to do `../../../../folder` just to navigate through the directory.
 
-![image](https://github.com/SPARCS-UP-Mindanao/SPARCS-Event-Platform/assets/85269524/08d2ed2c-84b7-4a93-a110-0bf5af71d4dc)
-
-## Styling guide
-
-Our project has a design system. You can view it [here](https://www.figma.com/file/F2HfagNiajlBmN4vDXvqg1/SPARCS-Ticket-%26-Certificate-System?type=design&node-id=0-1&mode=design&t=WFJeam4Qa4DSPi6F-0).
-
-Fortunately, the colors and fonts are now integrated tailwind, which makes it easy for us to style.
-![image](https://github.com/SPARCS-UP-Mindanao/SPARCS-Event-Platform/assets/85269524/801c200c-22b1-4762-afd2-65dbd3d547a9)
-
-colors and fonts can simply be accessed in these ways:
-
-- `className="text-primary-500"`
-- `className="bg-secondary-pink-300"`
-- `className="font-raleway"`
-- `className="font-subjectivity"`
-
 If you're new to tailwind, you can also read about tailwind here and the cheat sheet.
 
 - [Tailwind Docs](https://tailwindcss.com/)
@@ -58,24 +42,16 @@ If you're new to tailwind, you can also read about tailwind here and the cheat s
 
 Also, please install the [tailwind extension](https://marketplace.visualstudio.com/items?itemName=bradlc.vscode-tailwindcss) to have autocomplete feature.
 
-### Icons
+# Generating amplify_outputs.json and .env file
 
-In this project, we will be using the library [Phosphor Icons](https://phosphoricons.com/). The `Icon.tsx` component is already configured so that we won't have to import every icon.
+Utilize these scripts to generate the files required to run the project locally.
 
-![image](https://github.com/SPARCS-UP-Mindanao/SPARCS-Event-Platform/assets/85269524/bc6587de-04f0-4691-9d7b-e73b6ff0e415)
+Note: You must first setup AWS CLI in your machine and login with your provided AWS account. You might also need to do `export AWS_PROFILE=<your-profile>`.
 
-Take note of the component name. In this case, it's `CaretCircleRight`.
+$stage = dev | staging
 
-Here is how you can use the `Icon.tsx` component
-
-```
-  <Icon name="CaretCircleRight" weight="bold" className="h-4 w-4" />
-```
-
-Here are the other props that we can include in the Icon component
-![image](https://github.com/SPARCS-UP-Mindanao/SPARCS-Event-Platform/assets/85269524/1be66947-d8d6-4090-8d61-067f202f3d1a)
-
-In this project, we will be using `bold` and `fill` weight, where `bold` is the default value.
+- `npm run generate:env $stage`
+- `npm run generate:outputs $stage`
 
 ## Components
 
@@ -136,8 +112,9 @@ For naming hooks, we will use Camel case with the starting word `use`. Example: 
 ## Code formatting
 
 Please use `prettier` for formatting code to make it look cleaner.
+You can find it as an extension in `VSCode`.
 
-You can find it as an extension in `VSCode`
+You may also utilized the command `npm run pretty` to format the code.
 
 ##
