+++ frontend/package.json
@@ -64,7 +64,6 @@
     "react-router-dom": "^7.8.1",
     "tailwind-merge": "^3.3.1",
     "tailwindcss-animate": "^1.0.7",
-    "typescript-cookie": "^1.0.6",
     "ulid": "^3.0.1",
     "zod": "^4.0.17"
   },
