+++ frontend/scripts/generate-env.mjs
@@ -1,12 +1,13 @@
 import fs from 'fs';
 import path from 'path';
 import url from 'url';
-import { parseArgs } from './utils.mjs';
+import { parseArgs, mustBeUrl, chunk, isBad } from './utils.mjs';
 import { SSMClient, GetParametersCommand } from '@aws-sdk/client-ssm';
 
-// Tiny args: node scripts/generate-dotenv.mjs prod --region ap-southeast-1 --out .env
+// Tiny args: node scripts/generate-dotenv.mjs staging --region ap-southeast-1 --out .env
 
 const args = parseArgs(process.argv);
+
 const STAGE = args._[0] || process.env.SSM_STAGE; // dev | staging | prod
 const REGION = args.region || process.env.AWS_REGION || 'ap-southeast-1';
 const OUT = args.out || '.env';
@@ -25,38 +26,19 @@ const OUT_ABS = path.isAbsolute(OUT) ? OUT : path.join(__dirname, '..', OUT);
 
 const ssm = new SSMClient({ region: REGION });
 
-// Logical env keys -> SSM parameter names
-const PARAMS = {
+/**
+ * Map of logical keys -> array of candidate SSM parameter names (first found wins).
+ */
+const parameterNames = {
   VITE_API_AUTH_BASE_URL: `/techtix/auth-api-url-${STAGE}`,
   VITE_API_EVENTS_BASE_URL: `/techtix/events-api-url-${STAGE}`,
   VITE_API_PAYMENT_BASE_URL: `/techtix/payment-api-url-${STAGE}`
 };
 
-function chunk(a, n) {
-  const out = [];
-  for (let i = 0; i < a.length; i += n) out.push(a.slice(i, i + n));
-  return out;
-}
-
-function isBad(val) {
-  if (val == null) return 'missing';
-  const t = String(val).trim();
-  if (!t) return 'empty';
-  if (['undefined', 'null'].includes(t.toLowerCase())) return `invalid ('${val}')`;
-  return null;
-}
-
-function mustBeUrl(key, val) {
-  try {
-    new URL(val);
-  } catch {
-    throw new Error(`${key} is not a valid URL: ${val}`);
-  }
-}
-
 async function fetchParams(names) {
   const nameToValue = {};
-  const invalid = [];
+  const invalidParams = [];
+
   for (const batch of chunk(names, 10)) {
     const resp = await ssm.send(
       new GetParametersCommand({
@@ -65,9 +47,9 @@ async function fetchParams(names) {
       })
     );
     for (const p of resp.Parameters ?? []) nameToValue[p.Name] = p.Value;
-    if (resp.InvalidParameters?.length) invalid.push(...resp.InvalidParameters);
+    if (resp.InvalidParameters?.length) invalidParams.push(...resp.InvalidParameters);
   }
-  return { nameToValue, invalid };
+  return { nameToValue, invalid: invalidParams };
 }
 
 function serializeEnv(obj) {
@@ -82,13 +64,13 @@ function serializeEnv(obj) {
 async function main() {
   console.log(`ℹ️ Generating ${OUT} for stage="${STAGE}" (region: ${REGION})`);
 
-  const names = Object.values(PARAMS);
+  const names = Object.values(parameterNames);
   const { nameToValue, invalid } = await fetchParams(names);
 
   const missing = [];
   const envOut = {};
 
-  for (const [envKey, ssmName] of Object.entries(PARAMS)) {
+  for (const [envKey, ssmName] of Object.entries(parameterNames)) {
     const val = nameToValue[ssmName];
     const bad = isBad(val);
     if (bad) missing.push(`${envKey} (${ssmName}) ${bad}`);
