--- frontend/src/api/test.ts
@@ -1,7 +1,0 @@
-import { createApi } from './utils/createApi';
-
-export const getTest = (name: string) =>
-  createApi({
-    url: '/pokemon',
-    body: { name }
-  });
