+++ frontend/src/pages/admin/updatePassword/UpdatePasswordForm.tsx
@@ -1,10 +1,8 @@
 import { useNavigate } from 'react-router-dom';
 import { FormProvider } from 'react-hook-form';
-import { removeCookie } from 'typescript-cookie';
 import Button from '@/components/Button';
 import { FormItem, FormLabel, FormError } from '@/components/Form';
 import Input from '@/components/Input';
-import { cookieConfiguration } from '@/utils/cookies';
 import { useMetaData } from '@/hooks/useMetaData';
 import { useAdminUpdatePasswordForm } from '@/hooks/userAdminUpdatePasswordForm';
 
@@ -13,14 +11,9 @@ const UpdatePasswordForm = () => {
   setMetaData({});
   const navigate = useNavigate();
   const successHandler = () => {
-    removeCookie('_auth', cookieConfiguration);
-    removeCookie('_auth_user', cookieConfiguration);
-    removeCookie('_auth_refresh', cookieConfiguration);
-    removeCookie('_auth_storage', cookieConfiguration);
-    removeCookie('_auth_type', cookieConfiguration);
-    removeCookie('_auth_refresh_time', cookieConfiguration);
     navigate('/admin/login');
   };
+
   const { form, submit, isSubmitting } = useAdminUpdatePasswordForm(successHandler);
 
   return (
